buddy-django/buddyproject/requirements.txt
asgiref==3.10.0
distlib==0.3.9
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.18.0
platformdirs==4.3.8
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
